ZipExtensions now part of Utils -> can be used to create (un)zip-nodes for all dx11 types
