+++ Nodes/VVVV.DX11.Nodes/VVVV.DX11.Nodes.csproj
@@ -21,6 +21,10 @@
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="VVVV.Nodes.Generic, Version=32.0.0.0, Culture=neutral, processorArchitecture=x86">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\Dependencies\x86\VVVV.Nodes.Generic.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Nodes\Buffers\Lights\DirectionalLightBuffer.cs" />
@@ -74,6 +78,7 @@
     <Compile Include="Nodes\Textures\Array\GetSliceTextureArray.cs" />
     <Compile Include="Nodes\Textures\Array\TextureArraySetSlice.cs" />
     <Compile Include="Nodes\Textures\Array\TextureArraySetSliceNode.cs" />
+    <Compile Include="Nodes\Utils\ZipExtensions.cs" />
     <Compile Include="Nodes\Validators\RemoveSliceValidatorNode.cs" />
     <Compile Include="ShaderUtils.cs" />
     <EmbeddedResource Include="effects\quad.fx" />
